requirements.txt
openai==1.6.1
requests==2.31.0
aiohttp==3.8.1
pdf2image==1.16.0
Pillow==9.0.0